refactor : features-based-architecture 형식으로 맞춤. + 경로 단순화  패치
