+++ src/features/header/components/LoginMenu.jsx
@@ -1,0 +1,33 @@
+import React, { useContext } from 'react';
+import { useNavigate } from 'react-router-dom';
+import KakaoLoginBtn from '../../auth/components/KakaoLoginBtn';
+import useLogout from '../../auth/hooks/useLogout';
+import { AuthContext } from '@/contexts/AuthContext';
+export default function LoginMenu() {
+    const { user, setUser } = useContext(AuthContext);
+    const navigate = useNavigate();
+
+   const logout = useLogout();
+
+    return (
+        <div className="right-menu">
+            {user ? (
+                <>
+                    <img
+                        src={user.profile}
+                        alt="프로필"
+                        className="profile-image"
+                        onClick={() => navigate('/profile')}
+                    />
+                    <span className="nickname">{user.nickname}님, 환영합니다.</span>
+                    <button onClick={logout} className="logout-button">
+                        로그아웃
+                    </button>
+                </>
+            ) : (
+                <KakaoLoginBtn className="login-button" />
+            )}
+            <button className="more-button">⋮</button>
+        </div>
+    );
+}
