+++ vite.config.js
@@ -1,6 +1,12 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import path from 'path';
 
 export default defineConfig ({
 	plugins: [react()],
+	resolve: {
+		alias: {
+			'@' : path.resolve(__dirname, './src'),
+		},
+	},
 });
